Language Trivia Page Mold

+++ codecracker/User.swift
@@ -2,12 +2,12 @@
 //  User.swift
 //  codecracker
 //
-//  Created by user264588 on 9/1/24.
+//  Created by Gabriel Scholze on 9/1/24.
 //
 
 import Foundation
 
-struct User {
-    var java: Int
-    var cpp: Int
+struct User{
+    var java: Double
+    var cpp: Double
 }
